refactor Kernel class
