+++ .env
@@ -15,7 +15,9 @@
 
 ###> symfony/framework-bundle ###
 APP_ENV=dev
-APP_SECRET=c8d926cce77883722d85effa7315d362
+# APP_SECRET=c8d926cce77883722d85effa7315d362
+APP_HOST=localhost
+APP_PORT=80
 #TRUSTED_PROXIES=127.0.0.0/8,10.0.0.0/8,172.16.0.0/12,192.168.0.0/16
 #TRUSTED_HOSTS='^(localhost|example\.com)$'
 ###< symfony/framework-bundle ###
