+++ README.md
@@ -1,3 +1,3 @@
 **Symfony + ReactPHP simple Application**
 Get started:
-``php bin/console server:run --host='localhost' --posrt='80''``
+``php bin/console server:run --host=localhost --posrt=80``
